Add commands for running TEACh with EMMA (#7)
Make it easier to run TEACh with EMMA.

This is a **huge update** that improves the CLI massively. 

To know how to do things, just run `python -m emma_experience_hub teach` and you'll be able to see all the commands. 

This is an example of what you should see: 

<img width="866" alt="Screenshot 2022-08-19 at 16 35 44@2x" src="https://user-images.githubusercontent.com/7276308/185655150-0cf1a6a5-76c1-47da-8d1c-41c411610c58.png">



Things that _might_ need addressing: 

- [ ] `subprocess` is being used to download files from S3. While not dire, this is poor practice and we should switch to using Boto3 or something.
- [x] Download EDH instances
- [x] Create a subset of EDH instances
- [x] Calculate the metrics from the outputs
- [x] Run the AI2THOR environment and the game