Improve the SimBotAction datamodel with the new `payload` field (#43)
I've added a `payload` field into the `SimBotAction` and have updated
the validator to make the parsing a lot easier.

Essentially, we can access the field using the `.payload` attribute. But
when we use `.dict()` or `.json()`, the field name _should_ always be
what the SimBot Arena API will expect from us.