Generalise action parser to ignore wrong actions (#67)
Given the latest version of the model, we noticed that it was generating
some parts of the action trajectory wrong but the last part was actually
perfect. This means that the parser should also be able to support cases
in which some of the actions in the generated trajectory are faulty.
Example below:

```
 "prediction": "<s>g around. goto sink <frame_token_5>.</s>.",

  "groundtruth": " goto sink <frame_token_5>.</s>"
```

In this case, we want to parse `Goto sink <frame_token_5>` and ignore
the rest. This will still generate a reasonable prediction and work in
most cases.

Co-authored-by: Amit Parekh <7276308+amitkparekh@users.noreply.github.com>