feat: add simple feedback for errors from the arena (#131)
This also adds some more logs, and introduces some little hacks to make
sure we are sending the correct output to the policy. The hacks will be
fixed and removed in the future.

This also fixes the issue where EMMA _loves_ saying "Done!" to
everything.

This is also currently in production and verified working. 

Also, clarification questions have been disabled again because it was
causing the model to hallucinate. This is being looked into by @gpantaz.

I'm sorry for shoving multiple things into the same PR.