fix: improve how dialog intents are retrieved from previous actions (#190)
Bit of a chunky refactor here, but definitely needed. 

The `SimBotDialogPayload` now has an intent property for tracking the
`SimBotIntentType` for that utterance. But that caused circular imports
so the various enums are now in their own subpackage to prevent circular
imports.

There is also now a `SimBotDialogAction` to simplify `isinstance()`
checks when getting the intent from the payload.