fix: do not re-serialise the response model to `SimBotResponse` on returning endpoint (#196)
This was causing exceptions because the `SimBotDialogAction` requires
the `intent` field but it's not included when exporting to JSON.

## Why did this happen? 

The endpoint for the API tries to return a `SimBotResponse` model.
Within this response is the `actions` field, which is a
`list[SimBotAction]`. This list of actions comes directly from the
`SimBotSessionTurnActions.to_list()` method. On returning the endpoint,
FastAPI validates the `SimBotResponse` model by running all the
validators for it and all its fields. It appears to do this by exporting
the model to a JSON/dict and then re-serialising it as a
`SimBotResponse`.

To help make it easier to track the intents behind dialog, the field
`intent` was added to the `SimBotDialogPayload`. This field was required
and always expected. On exporting the model, the `intent` field is
explicitly excluded from the JSON so that we return _exactly_ what is
expected by the Arena.

When converting the response to JSON and then serializing it again, the
`SimBotAction` model attempts to detect and convert the payloads to the
correct type (see validators in `SimBotAction`). Since the `intent`
field does not exist in the JSON, it was causing the exception.