feat: include/support alias types for the Goto action (#203)
This is useful because we can more easily detect and track what goto
actions we have performed.

There's also been some refactoring of class methods in
`SimBotActionType` so that base types with aliases can be more easily
found/parsed.

Additionally, `SimBotAction` will now check if the action type is one of
the base types, and try to convert the alias type and the payload to its
most-specific variant. If it can't be done, it's just passed through as
is.

Since we control all the pieces, it should always be possible to find
the most-specific type and convert it. If we can't something dire has
clearly gone wrong and that's a whole different issue.