fix(deployment): add compound splitter to the docker compose (#211)
`main` is currently broken as the compound splitter service is not
included in the docker-compose or the Service Registry.

Waiting for: 

- https://github.com/emma-simbot/compound-splitter/pull/10
- https://github.com/emma-simbot/compound-splitter/pull/9
- https://github.com/emma-simbot/compound-splitter/pull/1

---

To prevent this in the future, we need a way of checking that any
updates to clients/services are valid and will load properly. I'll think
on the best way and make issues for this.