feat: improve observability for the API (#217)
Start tracing the different spans and create services to automatically
send things to OpenSearch.

Note: I know about the password in the YAML file. It's been specifically
scoped so that all it can do is write to OpenSearch, but it's still not
ideal. Fixing it would probably take longer so it can either be done, or
it can be done later.