fix: compound splitter no longer adds to queue if its not needed (#224)
The compound splitter was adding the utterance to the queue, even if it
was something simple. This fixes that and makes sure to only add the
compound utterances to the queue if there is more than one split, and
makes sure to replace the utterance for the current turn by popping from
the queue.