feat: implement search plan (#208)
We need to decide what to do with the search plan. This PR has the
option1 enabled and call at the modules for the option 2 is commented
out.

Option1 (currently implemented): Find an object only from your current
viewpoint using turn left
Option2: Make the agent visit one more viewpoint using turn left
Option3: Instead of using the turn left to do the look around. I know
that I said this is not feasible but apparently I am wrong. The
advantage of this over 1 is that it will be faster but the user will not
see what we see.
Option4: Basically option2 but with the look around. 


I personally believe that we dont need the viewpoints so that leaves us
with option1/option4 but I cant decide. Basically the trade-off between
the turn left/look around is that with the turn left the search routine
feels more natural as the user sees what we see but we sacrifise speed
that we gain when sending the 4 images at the same time to the model
with the look around

Co-authored-by: MalvinaNikandrou <nikandroumalvina@gmail.com>
Co-authored-by: Ubuntu <ubuntu@ip-172-77-44-116.ec2.internal>
Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>