feat: optimise the compress segmentation mask function further (#240)
So it seems that the previous compress mask function (prior to #239) was
just as inefficient as the Alexa one. Traces from the past 24hours have
shown that there is still an issue with the compression mask speed.

This PR fixes it properly, and there are a bunch of tests to make sure
that it is working in the exact same way as the other function.

The table below shows the performance improvements for a 300px^2 image. 

<img width="1315" alt="Screenshot 2022-12-24 at 20 11 24@2x"
src="https://user-images.githubusercontent.com/7276308/209449894-1ebcaa0b-2c7f-411c-afba-8f5207827e33.png">


---

- [ ] This PR **has** been tested in the staging environment.
- [ ] This PR altered the decision-making process of the bot and these
alternations have been included in the [Figma
diagram](https://www.figma.com/file/lHucvo4KDMaex98TfgYbNq/Agent-Decision-Making-Process-Implementation?node-id=0%3A1&t=mzKjVjRXdeBty4co-1).