feat: customise timeouts for all the clients from the settings (#277)
This is primarily used for the T2 Evaluation, but this allows for
customising the timeouts for each and every httpx request made through
the client, so that the evaluation doesn't crash.

It has been set to httpx's default of 5 seconds. 

---

- [ ] This PR **has** been tested in the staging environment.
- [ ] This PR altered the decision-making process of the bot and these
alternations have been included in the [Figma
diagram](https://www.figma.com/file/lHucvo4KDMaex98TfgYbNq/Agent-Decision-Making-Process-Implementation?node-id=0%3A1&t=mzKjVjRXdeBty4co-1).