fix: Enable search across viewpoints and reset of utterance queue (#263)
- Enable the `GreedyMaximumVertexCoverSearchPlanner`
- For compound instructions such as `Find the fridge and open it`, we
need to reset the utterance queue if the fridge is not found. This
happens when the find queue is empty and the final action is RotateLeft,
because we have already searched the frame we will end up in.
- Also, I increased the `distance_threshold` to 3 and tested with it.

---

- [x] This PR **has** been tested in the staging environment.
- [ ] This PR altered the decision-making process of the bot and these
alternations have been included in the [Figma
diagram](https://www.figma.com/file/lHucvo4KDMaex98TfgYbNq/Agent-Decision-Making-Process-Implementation?node-id=0%3A1&t=mzKjVjRXdeBty4co-1).