fix: Avoid infinite search-no match loop by forcing act_one_match (#324)
If the search routine finds the wrong object, it can lead to an infinite
loop where the NLU outputs act_no_match, the utterance is added to the
queue, search finds the wrong object, we take the utterance out of the
queue etc. Therefore, when it's time to execute the original utterance
after the search we should skip the NLU and assign act_no_match to the
verbal intent.

I am also resetting the find queue after a new utterance or environment
error because this can lead to weird behaviors.


---

- [x] This PR **has** been tested in the staging environment.
- [x] This PR altered the decision-making process of the bot and these
alterations have been included in the [Figma
diagram](https://www.figma.com/file/lHucvo4KDMaex98TfgYbNq/Agent-Decision-Making-Process-Implementation?node-id=0%3A1&t=mzKjVjRXdeBty4co-1).