fix: lightweight feedback state (#344)
There is a problem currently where we can select a lightweight dialogue
response when we actually dont want to. This results in a dialogue
history object that contains the original instruction of the user + the
lightweight dialogue response that we have selected which are then sent
to policy.

---

- [x] This PR **has** been tested in the staging environment.
- [ ] This PR altered the decision-making process of the bot and these
alterations have been included in the [Figma
diagram](https://www.figma.com/file/lHucvo4KDMaex98TfgYbNq/Agent-Decision-Making-Process-Implementation?node-id=0%3A1&t=mzKjVjRXdeBty4co-1).