feat: avoid generating action when using lightweight dialog with a end of trajectory action (#353)
With the inclusion of responses that are lightweight dialogs with the
actions that include a stop token, it meant that a new request came to
the experience hub and we were defaulting to `<act><one_match>`.

This then triggered feature extraction and action generation which meant
that the user was then waiting for an additional second for the bot to
say "Done!".

There is no problems with feature extraction or action generation as it
returns the correct token and everything works perfectly, but it's just
slow. As a result, we need to put something in that checks for this
situation and means that we just bypass the action generation and return
the dialog response.


---

- [ ] This PR **has** been tested in the staging environment.
- [ ] This PR altered the decision-making process of the bot and these
alterations have been included in the [Figma
diagram](https://www.figma.com/file/lHucvo4KDMaex98TfgYbNq/Agent-Decision-Making-Process-Implementation?node-id=0%3A1&t=mzKjVjRXdeBty4co-1).