feat: Implementation of Gfh symbolic (#355)
This introduces the a symbolic GFH implementation. Instead of sending up
to X previous frames and try to find the closest viewpoint for the
object that we are interested we keep in track of the objects (entities)
that we see.

An entity is obtained from the perception and is _sort of_ the object
class except some special cases (freeze ray monitor, freeze ray shelf
etc).

The memory is essentially a dictionary where keys are rooms and values
dictionaries. Each value dictionary has the entity as key and as value
the SimBotEntity which corresponds to the area & the closest viewpoint
this specific entity has been observed. We read and write to the memory
every time we extract features and interact with the environment
(placing or picking up stuff).

When an act no match has occured the user is asked if we want to find
the object - entity. With an affirmative response we first try to find a
viewpoint based on the ghf history then include viewpoints based on the
max vertex cover. The find pipeline remains the same.

---

- [x] This PR **has** been tested in the staging environment.
- [ ] This PR altered the decision-making process of the bot and these
alterations have been included in the [Figma
diagram](https://www.figma.com/file/lHucvo4KDMaex98TfgYbNq/Agent-Decision-Making-Process-Implementation?node-id=0%3A1&t=mzKjVjRXdeBty4co-1).

---------

Co-authored-by: MalvinaNikandrou <nikandroumalvina@gmail.com>