feat: improve robot arm + button, search returns object name (#368)
There are a couple of updates on the policy that we should push to
production:

1. The action prediction now has the same hack used in the nlu to handle
utterances like `push the red button` when the agent is in front of the
robotic arm.
2. The NLU now returns the object name in `act search`. This is already
handled by the nlu parser so there is no need to update it. The benefit
of this change is that it enables the ghf in case of `act search`.
Previously the ghf was only triggered on `act no match`

---

- [x] This PR **has** been tested in the staging environment.
- [ ] This PR altered the decision-making process of the bot and these
alterations have been included in the [Figma
diagram](https://www.figma.com/file/lHucvo4KDMaex98TfgYbNq/Agent-Decision-Making-Process-Implementation?node-id=0%3A1&t=mzKjVjRXdeBty4co-1).