feat: remove highlight from search (#397)
The highlight action before going to the object is more of a headache
than an actual feature. We added a dummy move forward action with a 0
magnitude to reset the position of the agent so that the mask of the
highlight in the previous frame and the mask of the goto in the next
frame are identical and refer to the same object.

The problem with that is that by resetting the camera we may be unable
to spot an object that we could actually detect it without resetting it.
This is especially critical in GFH; we recall that we saw an object in a
particular spot but that object is no longer visible since we resetted
the camera. This actually happened today when we tried to go to the red
shelf after turning on the red computer.

Furthermore, we have a fixed budget of actions that we can perform
before actually sending the control back to the user. The highlight
provides extra overhead with all these extra actions that are tricky to
get right (move forward with 0 magnitude + highlight).

At this point it is just better off without this action. Removing the
highlight makes the find routine faster and the end result is the same.

---

- [x] This PR **has** been tested in the staging environment.
- [x] This PR altered the decision-making process of the bot and these
alterations have been included in the [Figma
diagram](https://www.figma.com/file/lHucvo4KDMaex98TfgYbNq/Agent-Decision-Making-Process-Implementation?node-id=0%3A1&t=mzKjVjRXdeBty4co-1).