feat: disable goto shelf plans (#418)
<!-- Pull Request description goes here -->
There is a problem with how we have thought about the pre-plans for the
goto shelfs. Since we create the plans before actually going to that
object we may not be able to correctly distinguish between the red and
the blue shelf. This means that by the time we arrive to one of the
shelfs (red) we may be confusing it with the other one (blue) and
therefore prompting the user for a plan that is completely off. The
solution for now is to disable the goto+shelf plans as they are too
risky to keep. We are still keeping though the place + shelf

---

- [x] This PR **has** been tested in the staging environment.
- [ ] This PR altered the decision-making process of the bot and these
alterations have been included in the [Figma
diagram](https://www.figma.com/file/lHucvo4KDMaex98TfgYbNq/Agent-Decision-Making-Process-Implementation?node-id=0%3A1&t=mzKjVjRXdeBty4co-1).