fix: Search in the current room if we just got there (#427)
Think of the mission of defrosting the cake. We pick up the cake and say
"Go to the breakroom table". Once we get to the breakroom we are not
looking at the table so we say "I remember seeing a table in the
robotics lab. Do you want me to search in the robotics lab?"

---

- [ ] This PR **has** been tested in the staging environment.
- [ ] This PR altered the decision-making process of the bot and these
alterations have been included in the [Figma
diagram](https://www.figma.com/file/lHucvo4KDMaex98TfgYbNq/Agent-Decision-Making-Process-Implementation?node-id=0%3A1&t=mzKjVjRXdeBty4co-1).