feat: Handle ask_about_location based on memory (#527)
For ask_about_location intents, if the entity is seen before we just go
for it. If it is a pre-defined (large) object not in memory , we ask
whether to search. Otherwise, default to ask_about_locate response.

- [x] This PR **has** been tested in the staging environment.
- [ ] This PR altered the decision-making process of the bot and these
alterations have been included in the [Figma
diagram](https://www.figma.com/file/lHucvo4KDMaex98TfgYbNq/Agent-Decision-Making-Process-Implementation?node-id=0%3A1&t=mzKjVjRXdeBty4co-1).

---------

Co-authored-by: MalvinaNikandrou <nikandroumalvina@gmail.com>